refactor(maencof): reorganize hooks into utils and update import paths
- Moved various hooks into a new `utils` directory for better organization and clarity.
- Updated import paths across multiple test and source files to reflect the new structure.
- Introduced new utility functions for config provisioning, lifecycle dispatching, and vault committing.
- Enhanced documentation for new utility modules to clarify their purpose and usage.
- Ensured all tests are aligned with the updated import paths and structure.
