+++ plugins/maencof/src/hooks/postToolUse/postToolUse.entry.ts
@@ -6,7 +6,7 @@ import { writeResult } from '../shared/writeResult.js';
 import type {
   DispatchInput,
   MergedHookOutput,
-} from '../utils/dispatchTypes.js';
+} from '../../types/dispatch.js';
 
 import { orchestratePostToolUse } from './postToolUse.js';
