+++ plugins/maencof/src/hooks/postToolUse/postToolUse.ts
@@ -1,12 +1,12 @@
-import { runLifecycleDispatcher } from '../lifecycleDispatcher/lifecycleDispatcher.js';
+import { runLifecycleDispatcher } from '../utils/lifecycleDispatcher/lifecycleDispatcher.js';
 import { MAENCOF_MCP_TOOLS } from '../shared/maencofMcpTools.js';
 import type {
   DispatchInput,
   HookConcernResult,
   MergedHookOutput,
-} from '../utils/dispatchTypes.js';
-import { mergeHookOutput } from '../utils/mergeHookOutput.js';
-import { safeConcern } from '../utils/safeConcern.js';
+} from '../../types/dispatch.js';
+import { mergeHookOutput } from '../utils/mergeHookOutput/mergeHookOutput.js';
+import { safeConcern } from '../utils/safeConcern/safeConcern.js';
 
 import { runActivityRecorder } from './helpers/activityRecorder/activityRecorder.js';
