+++ plugins/maencof/src/hooks/preToolUse/preToolUse.entry.ts
@@ -6,7 +6,7 @@ import { writeResult } from '../shared/writeResult.js';
 import type {
   DispatchInput,
   MergedHookOutput,
-} from '../utils/dispatchTypes.js';
+} from '../../types/dispatch.js';
 
 import { orchestratePreToolUse } from './preToolUse.js';
