+++ plugins/maencof/src/hooks/sessionEnd/sessionEnd.entry.ts
@@ -6,7 +6,7 @@ import { writeResult } from '../shared/writeResult.js';
 import type {
   DispatchInput,
   MergedHookOutput,
-} from '../utils/dispatchTypes.js';
+} from '../../types/dispatch.js';
 
 import { orchestrateSessionEnd } from './sessionEnd.js';
