+++ plugins/maencof/src/hooks/sessionEnd/sessionEnd.ts
@@ -1,11 +1,11 @@
-import { runLifecycleDispatcher } from '../lifecycleDispatcher/lifecycleDispatcher.js';
+import { runLifecycleDispatcher } from '../utils/lifecycleDispatcher/lifecycleDispatcher.js';
 import type {
   DispatchInput,
   MergedHookOutput,
-} from '../utils/dispatchTypes.js';
-import { mergeHookOutput } from '../utils/mergeHookOutput.js';
-import { safeConcern, safeConcernAsync } from '../utils/safeConcern.js';
-import { runVaultCommitter } from '../vaultCommitter/vaultCommitter.js';
+} from '../../types/dispatch.js';
+import { mergeHookOutput } from '../utils/mergeHookOutput/mergeHookOutput.js';
+import { safeConcern, safeConcernAsync } from '../utils/safeConcern/safeConcern.js';
+import { runVaultCommitter } from '../utils/vaultCommitter/vaultCommitter.js';
 
 import { runSessionEnd } from './helpers/finalize/finalize.js';
