+++ plugins/maencof/src/hooks/utils/configProvisioner/INTENT.md
