+++ plugins/maencof/src/hooks/utils/configProvisioner/configProvisioner.ts
@@ -1,8 +1,8 @@
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'node:fs';
 
-import { appendErrorLogSafe } from '../../core/errorLog/errorLog.js';
+import { appendErrorLogSafe } from '../../../core/errorLog/errorLog.js';
 import { CONFIG_REGISTRY } from '../configRegistry/configRegistry.js';
-import { metaPath } from '../shared/metaPath.js';
+import { metaPath } from '../../shared/metaPath.js';
 
 export interface ProvisionResult {
   /** Filenames that were newly created */
