+++ plugins/maencof/src/hooks/utils/configRegistry/INTENT.md
