+++ plugins/maencof/src/hooks/utils/vaultCommitter/DETAIL.md
