+++ plugins/maencof/src/hooks/utils/vaultCommitter/INTENT.md
