+++ plugins/maencof/src/types/dispatch.ts
@@ -1,4 +1,4 @@
-import type { LifecycleEvent } from '../../types/lifecycle.js';
+import type { LifecycleEvent } from './lifecycle.js';
 
 /** The six Claude Code hook events maencof dispatches. */
 export type DispatchEvent = LifecycleEvent;
